fix:check bucket exist when delete bucket

+++ http/objectstore/s3api/s3_bucket_handler.go
@@ -103,14 +103,16 @@ func (s3a *s3ApiServer) DeleteBucketHandler(w http.ResponseWriter, r *http.Reque
 
 	bucket, _ := getBucketAndObject(r)
 	log.Infof("DeleteBucketHandler %s", bucket)
-
-	// avoid duplicated buckets
 	cred, _, err := s3a.authSys.CheckRequestAuthTypeCredential(context.Background(), r, s3action.DeleteBucketAction, bucket, "")
 	if err != api_errors.ErrNone {
 		response.WriteErrorResponse(w, r, err)
 		return
 	}
-
+	get := s3a.authSys.PolicySys.Head(bucket, cred.AccessKey)
+	if !get {
+		response.WriteErrorResponse(w, r, api_errors.ErrNoSuchBucketPolicy)
+		return
+	}
 	errc := s3a.authSys.PolicySys.Delete(context.Background(), cred.AccessKey, bucket)
 	if errc != nil {
 		response.WriteErrorResponse(w, r, api_errors.ErrNoSuchBucketPolicy)
